split config
