+++ backend/src/core/agent/prompt_util.py
@@ -8,7 +8,6 @@ from langchain_core.prompts import (
 )
 
 from ..lib_config import get_lib_config
-
 from ..prompt_mgr import list_prompts
 from ..public_models import AgentPromptStatus
 from .internal_models import AgentPromptName
