+++ backend/src/core/doc_mgr/upload/upload.py
@@ -2,7 +2,6 @@ import logging
 from pathlib import Path
 
 from ...lib_config import get_lib_config
-
 from ...providers.file_store import get_s3_bucket, get_s3_directory
 from ..doc.add import add_document
 from ..doc_set.query import get_document_sets, list_document_sets
