--- backend/src/core/lib_config/__init__.py
@@ -1,4 +1,0 @@
-from .main import get_lib_config
-
-# Explicitly define the exported names: these names are the contract of this module.
-__all__ = ['get_lib_config']
