+++ backend/src/core/providers/doc_loader/unstructured.py
@@ -6,7 +6,6 @@ import nltk
 from langchain_unstructured import UnstructuredLoader
 
 from ...lib_config import get_lib_config
-
 from ...signals import start_up_handler
