+++ backend/src/core_app/app_config.py
@@ -5,19 +5,9 @@ provides a global configuration object.
 
 from functools import cache
 from pathlib import Path
-from typing import Union
+import os
 
-from pydantic import (
-    AnyHttpUrl,
-    BaseModel,
-    DirectoryPath,
-    Field,
-    FilePath,
-    NewPath,
-    PostgresDsn,
-    RedisDsn,
-    StringConstraints,
-)
+from pydantic import Field, StringConstraints
 
 # See https://docs.pydantic.dev/latest/api/types/#pydantic.types.StringConstraints
 from pydantic_settings import BaseSettings, PydanticBaseSettingsSource, SettingsConfigDict, TomlConfigSettingsSource
@@ -31,7 +21,7 @@ LowerCaseStr = Annotated[str, StringConstraints(to_lower=True)]
 PasswordOrKeyStr = Annotated[str, StringConstraints(min_length=8)]
 
 
-class Settings(BaseSettings):
+class ApplicationSettings(BaseSettings):
     """
     Application-wide configuration settings loaded from environment variables and TOML files.
 
@@ -56,7 +46,11 @@ class Settings(BaseSettings):
         # We assume that the .env files were loaded into the environment
         # on an earlier step.
 
-        toml_file_path = Path('./config_data/app_data.toml')
+        toml_file_path = (
+            Path(env_var_value) 
+            if (env_var_value := os.environ.get('CONFIG_TOML_FILE')) 
+            else Path('./config.toml')
+        )
 
         # init_settings: setting values provided as keyword arguments when initialization
         #     an instance of this Settings class.
@@ -73,21 +67,17 @@ class Settings(BaseSettings):
             TomlConfigSettingsSource(settings_cls, toml_file=toml_file_path),
         )
 
+    jwt_write_claim_missing_ok: bool = Field(default=False, validation_alias='JWT_WRITE_CLAIM_MISSING_OK')
+
     application_jwt_secret: JwtSecretStr = Field(default='', validation_alias='APPLICATION_JWT_SECRET')
 
 
 @cache
-def get_auth_config():
+def get_app_config():
     """
-    Get the cached global application configuration instance.
+    Get the cached global application configuration.
 
-    Returns the singleton Settings object that contains all application
-    configuration loaded from environment variables, TOML files, and other sources.
-    Uses functools.cache to ensure the configuration is loaded only once per
-    application lifecycle for optimal performance.
-
-    Returns:
-        Settings: The application configuration object containing JWT secrets,
-                 database connections, API keys, and other runtime settings.
+    Returns a Settings instance loaded from environment variables and TOML files.
+    Configuration is cached for performance.
     """
-    return Settings()
+    return ApplicationSettings()
