+++ backend/src/core_app/routers/admin.py
@@ -5,8 +5,9 @@ from fastapi import APIRouter, Depends, Query
 
 from core.signals import send_reset_data
 from core_worker import reset_data_task
-from ..app_config import get_app_config
 from simple_auth import Scope, User, get_scoped_current_user
+
+from ..app_config import get_app_config
 
 logger = logging.getLogger(__name__)
