--- backend/tests/test_config/__init__.py
@@ -1,4 +1,0 @@
-from .main import get_test_config
-
-# Explicitly define the exported names: these names are the contract of this module.
-__all__ = ['get_test_config']
